验证BST

+++ src/main/java/tree/TreeNode.java
@@ -1,11 +1,11 @@
 package tree;
 
 public class TreeNode {
-    public Integer val;
+    public Long val;
     public TreeNode left;
     public TreeNode right;
 
-    public TreeNode(Integer val) {
+    public TreeNode(Long val) {
         this.val = val;
     }
 }
